maint: stop stubbing log level setting.
The underlying problem turned up when another test (as a side effect) changed
the logging level to be more verbose, and a very distant test broke.  This
revealed that we didn't preserve that global state around tests.

Fixing that further revealed that we stubbed setting that log level all over
the place, as a point fix for the same problem, and to assert the operation of
various tools.

So, additionally, we strip out all that stubbing, and assert on the desired
effect rather than the mechanism for achieving it.

Reviewed-By: Max Martin <max@puppetlabs.com>