(#7183) Implement "invisible glob" version matching for faces
"Invisible glob", or "prefix", version matching means that when you specify a
version string to use you can specify as little as one version number out of
the semantic versioning spec.

Matching is done on the prefix; an omitted number is treated as "anything" in
that slot, and we return the highest matching versioned face by that spec.

For example, given the set of versions: 1.0.0, 1.0.1, 1.1.0, 1.1.1, 2.0.0
The following would be matched:

  input    matched
  1        1.1.1
  1.0      1.0.1
  1.0.1    1.0.1
  1.0.2    fail - no match
  1.1      1.1.1
  1.1.1    1.1.1
  1.2      fail - no match