maint: fix a race in catalog compilation versioning.
The implementation has a whole pile of time dependencies in our comparisons,
we had a whole pile of races across the code.

We could try and fix the comparisons to work better, but that is actually
harder than it sounds thanks to the architecture.

Instead, freeze time for each test, ensuring that we consistently get the
result expected.