maint: add an "exit was called" matcher for rspec.
We have a bunch of places that want to test if exit was called in a block of
code or not, which we did in a whole pile of ad-hoc ways.  Instead, better to
have a custom matcher that tests specifically for the correct exit exception
being thrown, and the right error code in it.

Reviewed-By: Jesse Wolf <jesse@puppetlabs.com>