(#7101) Fix template error messages in Ruby 1.8.5
lib/puppet/parser/templatewrapper.rb#script_line was calling .first on a
String object, which in Ruby > 1.8.5 will return the entire string, but
in 1.8.5 will cause an exception. This change (proposed by Markus
Roberts) removes the call to .first and adds a condition for when the
template error involves the file not being found.

Reviewed-by: Jesse Wolfe