(#7279) Adding some basic file actions
Add the ability to download a file into the local filebucket using a puppet
URI or from disk.  Also, the ability to store into the filebucket.

These provide at least basic UI for moving data around using the filebucket
service, and act as an example of how further work can be done.

Also, update the code to eliminate a couple of redundant checks on arguments,
and add some basic testing around the actions.  Mostly only that they exist,
at this point.

Reviewed-By: Daniel Pittman <daniel@puppetlabs.com>