(#7248) Fail if two aliases of one option are passed...
From the command line, and other facades, it is fairly difficult to call an
action with two aliases of the same option in the invocation, but from the
Ruby API it would be relatively easy to achieve.

This is now checked, and raises an error, so that we don't accidentally have
strange or unusual behaviour coming out of the whole system.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>