(#7221) Strip bad whitespace from face and action docs.
We now strip whitespace in face (and related) documentation in two places:

We strip any trailing whitespace on each line, just because.

We strip any leading indent, but not all leading whitespace, from the text.
That is, we strip the *minimum* amount of whitespace that we can take from
every line in the documentation without changing the overall content.

Reviewed-By: Pieter van de Bruggen <pieter@puppetlabs.com>