(#7317) better error handling in CLI face facade.
The error handling code in the face CLI facade was pretty terrible: it make a
bunch of assumptions about the structure of the error message it got back, and
used exceptions to communicate inside the same function.

Instead, we handle errors uniformly without raising and catching in the same
method, report more nicely, and exit cleanly in all cases.

Reviewed-By: Nick Lewis <nick@puppetlabs.com>