(#7314) Faces fail horribly when one has a syntax error.
When we hit a syntax error in any face, a whole bunch of unrelated face things
would blow up in horrible ways.  Stack traces for all...

Now, instead, we catch that fault   but specifically only in the face file
and report it through our error logs, then quietly ignore the face.

Reviewed-By: Nick Lewis <nick@puppetlabs.com>