(#7329) Consistent naming for rendering formats and hooks.
We refer to rendering formats pretty consistently as `json`, `yaml`, `s`, and
so forth; unqualified names.

On the other hand, we refer to the rendering hooks *mostly* as `to_*`, except
the `:for_humans` and `:json` formats.  Which is kind of confusing because of
the internal inconsistency, and because it doesn't match the public name.

Fix the code to resolve both, so the `to_*` format still works, but we mostly
expect to use the `*` version, to match public expectation.