(#7353) Add 'console' format to FormatHandler
This adds a console rendering format to the Network FormatHandler subsystem;
it provides the same human-friendly textual rendering as the Faces application
did, except it uses JSON rather than PP as the fall-back rendering mode.

This paves the path for unification of all formatting into the same subsystem,
rather than the half-measures we used to have.

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>