(#7353) Unify rendering in the face_bace application.
We now move over to using only network FormatHandler rendering code for
output, ditching all the support we had in the application.

We include a compatibility shim to ensure that the :for_humans format that was
supported for a while is now an alias for the :console format we are using
moving forward.

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>