(#7291) The 'script' version of actions needs options argument
Ruby 1.9 is stricter about arity for for arguments passed as a block.  A
test case just hadn't been updated to take this into account for Face
scripts (a simpler form of action), which isn't surprising since script
isn't used anywhere in the code, which makes it a prime candidate for
removal as far as I'm concerned, but apparently Luke wants it in there.

Reviewed-by: Pieter van de Bruggen <pieter@puppetlabs.com>