(#7291) Fix Ruby 1.9 face failures
The certificate face wasn't being loaded, but it wasn't clear from the
test failure why:

    lib/puppet/interface.rb:61:in `[]': Could not find Puppet Face
    :certificate (Puppet::Error)

The problem is that when the certificate face is required you get:

    SyntaxError Exception:
    /Users/matthewrobinson/work/puppet/lib/puppet/face/certificate.rb:11:
    invalid multibyte char (US-ASCII)

However this error is caught and logged, but then ignored.  This
behavior was a decision in #7314 and is currently under review.

A space character in the description was ASCII 160 instead of the typical ASCII 32

Reviewed-by: Pieter van de Bruggen <pieter@puppetlabs.com>