(#7259) Do not try to load all Terminus classes when configuring the Indirector
When configuring the Indirector routes, we should only try loading the
Terminus classes that are referenced by the configuration.

Previously, we were loading all Terminus classes, which would cause
errors if we didn't have all of the prerequisites for all of them,
even if the ones with missing prerequisites weren't being used by the
configuration.

Paired-with: Nick Lewis <nick@puppetlabs.com>