(#7259) Remove ActiveRecord requirement from indirector face spec
"should be able to return a list of terminuses for a given
indirection" was calling Puppet::Indirector::Face.terminus_classes,
which is just a thin wrapper around
Puppet::Indirector::Terminus.terminus_classes, which would attempt to
load all Terminus classes.  This would cause problems if not all of
the prerequisites for all of the Terminus classes were installed (For
example: ActiveRecord).

Now we only test that the thin wrapper appropriately munges the output
from Puppet::Indirector::Terminus.terminus_classes, since the method
being wrapped should have its own tests for the behavior that was
being tested originally.

Paired-with: Nick Lewis <nick@puppetlabs.com>