(#7291) Fix issues with instance_methods in Ruby 1.9
instance_methods in Ruby 1.8.7 returns an array of strings, but returns
an array of symbols in 1.9.2.  This manifested itself when running the
tests because in 1.9.2 we were trying to call sub on a sybmol.  The
original proposed solution was to monkey patch symbols to have a sub
method, but this didn't deal with the real issue of need to check
whether a method was defined, and actually made it worse.

Turns out that checking for the presence of a method in an array that
may contain symbols and may contain strings is better done by just
calling method_defined? instead.

This patch addresses all the places ack turned up the code doing this
include? check instead of directly calling method_defined?.

Thanks to Alex Sharp ajsharp@gmail.com for pointing out the Ruby 1.9
problems and working toward a solution.

Reviewed-by: Nick Lewis <nick@puppetlabs.com>