(#7114) Fix specs for ssh authorized key parsed provider
These tests were unnecessarily creating a stub resource, as well as failing to
re-initvars on the provider before each test. This led to test ordering
failures, and incompatibilities with the other commits in this series. Now we
use a real resource, and properly reinitialize before each test.

Paired-With: Josh Cooper