(#7297) Fix Puppet::Resource#to_manifest in Ruby 1.9
This method was relying on the implicit join in Ruby 1.8's Array#to_s, eg.
[1,2,3].to_s => "123". The behavior in Ruby 1.9 is more akin to Array#inspect,
eg. [1,2,3].to_s => "[1, 2, 3]". Since the array we were building was lines
to be printed, the latter behavior is incorrect. So we just join into a
single string, which prints consistently in all versions of Ruby.

Paired-With: Josh Cooper
Original patch by Aria Stewart <aredridel@nbtsc.org>