maint: move trap call to Signal so we can stub it for specs
Also removed some monkey patching on Signal that would have
theoretically done this without having to explicitly call trap on Signal
in order to stub it, but it's not working.

This allows us to ctrl+c (send SIGINT) in the middle of a spec run.

Paired-with: Josh Cooper <josh@puppetlabs.com>