(#7624) Auditing should not be enabled by default for purged resources.
- Auditing creates logging messages as of 2.6.5 so it should not
  be enabled by default.

- This patch removes the :audit => :all settting from resources
  created via self.instances (which is used for purging).

Please note that we believe this change to be safe, and *should* not result in
user-visible behavioural differences when you use the `instances` method on a
type, but we can't give you a perfect assurance of that.

If you do have code that depends on the current behaviour, and it misbehaves
after this patch, please let us know so we can weep ^W find another solution
that works for everyone.

Reviewed-By: Daniel Pittman <daniel@puppetlabs.com>
Reviewed-By: Nigel Kersten <nigel@puppetlabs.com>