(#7632) Make secret_agent application compatible with secret_agent face
Running `puppet secret_agent` was failing with an error in validate_args (in
interface/action.rb), because the application was trying to pass the certname as
an argument to the synchronize action. Also, it was trying to submit the report
a second time. Reviewing the code with Nick0, we found that the application
wasn't inheriting from FaceBase and was duplicating a lot of work, and were able
to resolve the issue by basically deleting the whole thing.

This patch makes secret_agent behave like the other Faces apps, and makes
synchronize the default action of the secret_agent face. We left the `run_mode
:agent` line in the application because of bug #7802.

Paired-with: Nick Lewis <nick@puppetlabs.com>