(#7681) Allow array variables as resource references
The commit df088c9ba16dce50c17a79920c1ac186db67b9e9 introduced a regression
where

  $files = ["/tmp/one", "/tmp/two"]
  file { "/tmp/one": content => "one", }
  file { "/tmp/two": content => "two", }
  file { "/tmp/three":
    content => "three",
    require => File[$files],
  }

no longer worked.  File[$files] was concatenating the elements of $files to
create a single title, instead of expanding to multiple File dependencies.

Since resource reference titles are implicitly wrapped in an array, if one of
the elements of that array is a variable containing an array, the list of
titles is a nested array. Prior to the change causing the regression, we would
flatten arrays when evaluating them, under certain circumstances.  We no longer
ever flatten AST arrays when evaluating them, so anywhere that we really do
need a flattened array, we have to manually flatten it.

ResourceReference expects its list of titles to be a single, flat list of
titles, so we have to make it so.

Paired-with: Jacob Helwig <jacob@puppetlabs.com>