maint: Dedup the loadpath so we don't have to walk it multiple times
If the user's path has duplicate entries, we end up looking at them
multiple times.  This has bitten people in the past in that if you get a
lot of duplication it can make autloading a lot slower.  Really the user
shouldn't be duplicating their path, but since we can't control that,
this is a safe fix to prevent them from having autoload slowness.

Paired-with: Max Martin <max@puppetlabs.com>