(#7690) Don't blow up when listing terminuses available for faces
Previously, in order to list the available terminuses for an indirected
face we loaded all the the terminuses in order to list them.  This meant
that if a terminus like active_record didn't have the dependencies it
needed, the documentation would raise errors and not list terminuses.

    <Puppet::Error: Could not autoload filename uninitialized constant Object::ActiveRecord>

Now we just list the terminuses available in the load path without
trying to load them.  The terminus will still raise an error if you try
to use it without its dependencies being met.

Paired-with: Max Martin <max@puppetlabs.com>