maint: Fix order dependent spec failure for face indirection
An indirected face is being created in
spec/unit/application/indirection_base_spec.rb that uses a stubbed out
indirection.  This stub didn't have a name method defined, which caused
the documentation_spec.rb, that does checks against every available
face, to blow up with:

    expected no Exception, got #<NoMethodError: undefined method `to_sym' for nil:NilClass>

in lib/puppet/face/indirector/face.rb when listing the indirections for
the help text.

I toyed with creating a real indirection for the test, but that was
harder than expected.

Paired-with: Max Martin <max@puppetlabs.com>