(#7708) Delete extended documentation from configuration reference
The auto-generated references are meant to be pithy, dense, and fast -- to be
references, not guides. The configuration reference was front-loaded with
several pages of explanatory text that properly belongs in the guides on the
docs site. This commit removes that text in preparation for a reorganization of
the docs.

This is a doc string only commit.