(#7177) Deprecate implicit 'puppet apply' for 2.7.0
Back in prehistory (eg: 0.25 era), 'puppet' was the name for the agent, and
could be used directly to apply a manifest as well as to communicate with the
puppet master process.

During the 2.6 series we moved to a single binary, but continued to support
older scripts by detecting invocations that looked like the traditional
scripting uses and implicitly turning those into a call to 'puppet apply'.

Now, with the 2.7.0 release, we are moving to deprecate that behaviour.  We
still do the same detection, and still run the old manifests, but we now
emit a deprecation warning directing people to use 'puppet apply' directly.

We intend to remove the behaviour entirely in the 2.8 release, which also
paves the way to nicer handling of the command line.

Reviewed-By: Randall Hansen <randall@puppetlabs.com>
Reviewed-By: Nick Fagerlund <nick.fagerlund@puppetlabs.com>