(#7699) Don't duplicate inherited action names on faces.
We earlier moved to duplicating Action objects in the Faces subsystem to
ensure they had the correct binding context during execution and introspection.

This was correct, but introduced a bug where we would report both the parent
and child binding as separate entries with duplicate names, in the list of
actions.

This flowed on to the help output, where it would cause every inherited action
to be listed twice: once on the parent, once on the child.  (This was actually
worse if the inheritance was deeper: we would duplicate once for every level
between the instance and the origin of the action.)