maint: remove obsolete work-around code from help face.
The application wrapper for help used to disable the inherited 'render' method
from FaceBase - specifically, to avoid the behaviour of rendering strings
poorly.

Now that we have support for good output in the upstream method, this is
unnecessary, so we can eliminate the stub method entirely and use the default
behaviour.

(This also enables rendering the help into JSON or YAML, against the odds that
 someone actually cares about that. ;)