(#7683) Use ronn, when available, to render the output.
We now look for ronn(1), and if it is available ask it to generate the *roff
output, delegate that to man(1) directly, and show the result to the user in
their pager.

If ronn(1) isn't available we delegate to $MANPAGER, $PAGER, less, most, or
more, in that order, to paginate the raw markdown.  Not nearly so nice, but
better than doing nothing.

Finally, if none of those pagers are available we fall through to the default
behaviour of puppet rendering the output, which more or less results in a
direct dump to the console.  Nice.

Reviewed-By: Nick Fagerlund <nick.fagerlund@puppetlabs.com>