(#2128) Add the ability to specify a node name
The setting node_name_value may now be used for 'puppet apply' or 'puppet
agent' to specify the name for the node. This will not affect the certificate
used by the node, and the node will still be authenticated based on its
certname. The default value for node_name_value is the certname.

This is useful for eg. EC2 nodes whose random hostnames cannot be easily used
to classify them.

Paired-With: Jacob Helwig