(#2128) Get facts before retrieving catalog
Retrieving a catalog and getting the facts to submit with the catalog request
are distinct operations, and should be done separately. This is also to prepare
for adding the ability to determine the node name based on a fact, in which
case the node name needs to be determined before it is used for either the
catalog or the report.

Paired-With: Jacob Helwig <jacob@puppetlabs.com>