(#7624) Manually fetch all properties in instances.
When we removed the `:audit => :all` flag on creation of an instance from a
Puppet type, we stopped fetching all the properties.  This had flow-on
consequences that were visible from the outside; while some places did their
own work to ensure that properties were fetched, others didn't.

We now open-code the loop that creates and fetches those properties, to ensure
that we have the same data without going through the :audit machinery.

This resolves the excessive logging, and also eliminates the behavioural
change that we introduced in the previous commit.

Reviewed-By: Daniel Pittman <daniel@puppetlabs.com>