(#7728) Suppress notifications from container whits.
We introduced changes to our graph to reduce the number of edges when we had
container dependencies.  As part of this the 'whit' object was created, and
used as an intermediate vertex to simplify processing.

During that change we introduced additional reporting, based on those whit
objects, about containment relationships during resource application failure.

Specifically, we would now report that the containing class(es) and stages of
any failed resource were failed, point-blank.  This was unclear, because the
entire class had not failed, but only part of it, but also unhelpful, because
it turned a single failure into at least one additional report that contained
no additional information.

Now, instead, we suppress reporting for the whit resources.  We still process
them identically; just the report is eliminated.

It isn't absolutely clear that this is the correct long term direction for
handling these objects, but it is the minimal change for the RC release.

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>