maint: handle incoherent Test::Unit assertions.
So, it turns out that the 1.8.7 Test::Unit assertions are a bit quirky.  If we
use 'assert_match' it will turn a String into a regular expression by passing
it through Regexp.quote automatically.

If you use 'assert_no_match', though, it will not do that.  It will just
complain that it isn't a Regexp and blow up.  So, we manually quote outside
the assertion in both cases, and things should just work.