(#7828) Fix whitespace in synopsis generator.
We accidentally omitted whitespace between multiple options while building the
synopsis.  This fixes that, by introducing a breakable space in the right
location.

Additionally, we extract the code that was 99 percent identical from the face
and action synopsis generators, push it down into the documentation module,
and then invoke it from both places.

This eliminates the duplicate code, allowing me to fix that bug once and have
it apply to both parts of the code; this is pretty much assured to be true
any time we change the synopsis generation.

Reviewed-By: Nick Fagerlund <nick.fagerlund@puppetlabs.com>