(#2128) In-line docs for node_name_{fact,value}
This change augments the in-line documentation for the node_name_fact and
node_name_value configuration settings.  These settings will not work
effectively without additional changes elsewhere in the system, e.g. to
auth.conf.

In order to help the end user land softly if they choose to change these
settings, a short link URL we control and can redirect has been added
to each setting.  These currently point to the community Wiki but may be
redirected to docs.puppetlabs.com in the future.