(#2128) Whitespace only reflow commit
The output of puppet --genconfig was showing a blank line between the
description of the option and the default value line.  This is because
of the here document being used in the first commit.

This change replaces the HERE document with a normal string that does
not contain a final newline.

In addition, the indentation of one of the default options was using 4
spaces instead of 2 and this has been corrected.