(#7117) Use a different auth.conf instead of overwriting the default
This test was previously replacing the default auth.conf with its own, and
restoring it afterward. This was error-prone, failing if there wasn't an
initial auth.conf present. It would also causing cascading failures when this
test failed, by leaving a non-standard auth.conf for other tests to use. Since
this copy-and-restore behavior can be replaced by simply specifying a different
auth file to use, we do that instead.

Reviewed-By: Jacob Helwig <jacob@puppetlabs.com>
Reviewed-By: Dominic Maraglia <dominic@puppetlabs.com>